refactor: splitted predefined tasks into files add optional calls
