+++ src/rai_bench/rai_bench/tool_calling_agent/predefined/tasks.py
@@ -13,775 +13,18 @@
 # limitations under the License.
 
 import random
-from typing import List, Literal, Sequence
-
-from rai.tools.ros2 import MoveToPointToolInput
+from typing import List, Literal
 
 from rai_bench.tool_calling_agent.interfaces import (
     Task,
-    TaskArgs,
 )
-from rai_bench.tool_calling_agent.subtasks import (
-    CheckActionFieldsToolCallSubTask,
-    CheckArgsToolCallSubTask,
-    CheckTopicFieldsToolCallSubTask,
+from rai_bench.tool_calling_agent.predefined import (
+    get_basic_tasks,
+    get_custom_interfaces_tasks,
+    get_manipulation_tasks,
+    get_navigation_tasks,
+    get_spatial_tasks,
 )
-from rai_bench.tool_calling_agent.tasks.basic import (
-    AssessSensorDataQualityTask,
-    CheckRobotHealthTask,
-    GetAllROS2CamerasTask,
-    GetPointcloudTask,
-    GetRobotDescriptionTask,
-    GetROS2DepthCameraTask,
-    GetROS2RGBCameraTask,
-    GetROS2TopicsTask,
-)
-from rai_bench.tool_calling_agent.tasks.custom_interfaces import (
-    PublishROS2HRIMessageTextTask,
-)
-from rai_bench.tool_calling_agent.tasks.manipulation import (
-    MoveToPointTask,
-)
-from rai_bench.tool_calling_agent.tasks.navigation import (
-    MoveToBedTask,
-    MoveToFrontTask,
-    NavigateToPointTask,
-    SpinAroundTask,
-)
-from rai_bench.tool_calling_agent.tasks.spatial import (
-    BoolImageTaskEasy,
-    BoolImageTaskHard,
-    BoolImageTaskInput,
-    BoolImageTaskMedium,
-)
-from rai_bench.tool_calling_agent.validators import (
-    NotOrderedCallsValidator,
-    OrderedCallsValidator,
-)
-
-IMG_PATH = "src/rai_bench/rai_bench/tool_calling_agent/predefined/images/"
-true_response_inputs: List[BoolImageTaskInput] = [
-    BoolImageTaskInput(
-        question="Is the door on the left from the desk?",
-        images_paths=[IMG_PATH + "image_1.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is the light on in the room?",
-        images_paths=[IMG_PATH + "image_2.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Do you see the plant?",
-        images_paths=[IMG_PATH + "image_2.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Are there any pictures on the wall?",
-        images_paths=[IMG_PATH + "image_3.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Are there 3 pictures on the wall?",
-        images_paths=[IMG_PATH + "image_4.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is there a plant behind the rack?",
-        images_paths=[IMG_PATH + "image_5.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is there a pillow on the armchain?",
-        images_paths=[IMG_PATH + "image_7.jpg"],
-    ),
-]
-false_response_inputs: List[BoolImageTaskInput] = [
-    BoolImageTaskInput(
-        question="Is the door open?",
-        images_paths=[IMG_PATH + "image_1.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is someone in the room?",
-        images_paths=[IMG_PATH + "image_1.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Do you see the plant?",
-        images_paths=[IMG_PATH + "image_3.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Are there 4 pictures on the wall?",
-        images_paths=[IMG_PATH + "image_4.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is there a rack on the left from the sofa?",
-        images_paths=[IMG_PATH + "image_4.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is there a plant on the right from the window?",
-        images_paths=[IMG_PATH + "image_6.jpg"],
-    ),
-    BoolImageTaskInput(
-        question="Is there a red pillow on the armchair?",
-        images_paths=[IMG_PATH + "image_7.jpg"],
-    ),
-]
-########## SUBTASKS #######################################################################################
-get_topics_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_topics_names_and_types", expected_args={}
-)
-
-color_image5_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_image",
-    expected_args={"topic": "/color_image5"},
-    expected_optional_args={"timeout_sec": int},
-)
-depth_image5_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_image",
-    expected_args={"topic": "/depth_image5"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-color_camera_info5_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_image",
-    expected_args={"topic": "/color_image5"},
-    expected_optional_args={"timeout_sec": int},
-)
-depth_camera_info5_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_image",
-    expected_args={"topic": "/depth_image5"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-receive_robot_desc_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-receive_pointcloud_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/pointcloud"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# System health subtasks
-diagnostics_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/diagnostics"},
-    expected_optional_args={"timeout_sec": int},
-)
-rosout_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/rosout"},
-    expected_optional_args={"timeout_sec": int},
-)
-joint_states_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/joint_states"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Odometry subtasks
-odom_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/odom"},
-    expected_optional_args={"timeout_sec": int},
-)
-filtered_odom_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/odometry/filtered"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Transform subtasks
-tf_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/tf"},
-    expected_optional_args={"timeout_sec": int},
-)
-tf_static_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/tf_static"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-
-# Robot description subtasks
-robot_description_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description"},
-    expected_optional_args={"timeout_sec": int},
-)
-robot_description_semantic_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description_semantic"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Sensor data subtasks
-scan_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/scan"},
-    expected_optional_args={"timeout_sec": int},
-)
-pointcloud_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/pointcloud"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-
-# Robot description subtasks
-robot_description_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description"},
-    expected_optional_args={"timeout_sec": int},
-)
-robot_description_semantic_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description_semantic"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Sensor data subtasks
-scan_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/scan"},
-    expected_optional_args={"timeout_sec": int},
-)
-pointcloud_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/pointcloud"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Robot description subtasks
-robot_description_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description"},
-    expected_optional_args={"timeout_sec": int},
-)
-robot_description_semantic_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/robot_description_semantic"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-# Sensor data subtasks
-scan_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/scan"},
-    expected_optional_args={"timeout_sec": int},
-)
-pointcloud_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="receive_ros2_message",
-    expected_args={"topic": "/pointcloud"},
-    expected_optional_args={"timeout_sec": int},
-)
-
-
-######### MANIPULATION ########################################################################################
-move_to_point_subtask_grab = CheckArgsToolCallSubTask(
-    expected_tool_name="move_to_point",
-    expected_args={"x": 1.0, "y": 2.0, "z": 3.0, "task": "grab"},
-)
-move_to_point_subtask_drop = CheckArgsToolCallSubTask(
-    expected_tool_name="move_to_point",
-    expected_args={"x": 1.2, "y": 2.3, "z": 3.4, "task": "drop"},
-)
-
-pub_HRIMessage_text_subtask = CheckTopicFieldsToolCallSubTask(
-    expected_tool_name="publish_ros2_message",
-    expected_topic="/to_human",
-    expected_message_type="rai_interfaces/msg/HRIMessage",
-    expected_fields={"text": "Hello!"},
-)
-
-get_tohuman_interface_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="get_ros2_message_interface",
-    expected_args={"msg_type": "rai_interfaces/msg/HRIMessage"},
-)
-
-
-return_true_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="return_bool_response", expected_args={"response": True}
-)
-return_false_subtask = CheckArgsToolCallSubTask(
-    expected_tool_name="return_bool_response", expected_args={"response": False}
-)
-
-start_nav_action_subtask = CheckActionFieldsToolCallSubTask(
-    expected_tool_name="start_ros2_action",
-    expected_action="/navigate_to_pose",
-    expected_action_type="nav2_msgs/action/NavigateToPose",
-    expected_fields={
-        "pose": {
-            "header": {"frame_id": "map"},
-            "pose": {
-                "position": {"x": 2.0, "y": 2.0, "z": 0.0},
-            },
-        },
-    },
-)
-start_spin_action_subtask = CheckActionFieldsToolCallSubTask(
-    expected_tool_name="start_ros2_action",
-    expected_action="/spin",
-    expected_action_type="nav2_msgs/action/Spin",
-    expected_fields={"target_yaw": 3},
-)
-start_move_front_action_subtask = CheckActionFieldsToolCallSubTask(
-    expected_tool_name="start_ros2_action",
-    expected_action="/drive_on_heading",
-    expected_action_type="nav2_msgs/action/DriveOnHeading",
-    expected_fields={
-        "target": {"y": 0.0, "z": 0.0},
-    },
-)
-######### VALIDATORS #########################################################################################
-######### BASIC    ########################################################################################
-topics_ord_val = OrderedCallsValidator(subtasks=[get_topics_subtask])
-
-color_image_ord_val = OrderedCallsValidator(subtasks=[color_image5_subtask])
-depth_image_ord_val = OrderedCallsValidator(subtasks=[depth_image5_subtask])
-
-color_camera_info_ord_val = OrderedCallsValidator(subtasks=[color_camera_info5_subtask])
-depth_camera_info_ord_val = OrderedCallsValidator(subtasks=[depth_camera_info5_subtask])
-
-color_image_with_info_ord_val = NotOrderedCallsValidator(
-    subtasks=[color_image5_subtask, color_camera_info5_subtask]
-)
-depth_image_with_info_ord_val = NotOrderedCallsValidator(
-    subtasks=[depth_image5_subtask, color_camera_info5_subtask]
-)
-
-all_camera_images_notord_val = NotOrderedCallsValidator(
-    subtasks=[
-        color_image5_subtask,
-        depth_image5_subtask,
-    ]
-)
-all_camera_info_notord_val = NotOrderedCallsValidator(
-    subtasks=[
-        color_camera_info5_subtask,
-        depth_camera_info5_subtask,
-    ]
-)
-all_camera_images_with_info_notord_val = NotOrderedCallsValidator(
-    subtasks=[
-        color_image5_subtask,
-        depth_image5_subtask,
-        color_camera_info5_subtask,
-        depth_camera_info5_subtask,
-    ]
-)
-
-joint_states_ord_val = OrderedCallsValidator(subtasks=[joint_states_subtask])
-diagnostics_ord_val = OrderedCallsValidator(subtasks=[diagnostics_subtask])
-
-get_pointcloud_ord_val = OrderedCallsValidator(subtasks=[receive_pointcloud_subtask])
-get_robot_desc_ord_val = OrderedCallsValidator(subtasks=[receive_robot_desc_subtask])
-
-diagnostics_ord_val = NotOrderedCallsValidator(subtasks=[diagnostics_subtask])
-joint_states_ord_val = NotOrderedCallsValidator(subtasks=[joint_states_subtask])
-rosout_ord_val = NotOrderedCallsValidator(subtasks=[rosout_subtask])
-robot_health_val = NotOrderedCallsValidator(
-    subtasks=[diagnostics_subtask, joint_states_subtask, rosout_subtask]
-)
-
-odometry_comparison_val = NotOrderedCallsValidator(
-    subtasks=[odom_subtask, filtered_odom_subtask]
-)
-sensor_data_val = NotOrderedCallsValidator(
-    subtasks=[
-        scan_subtask,
-        pointcloud_subtask,
-        color_image5_subtask,
-        depth_image5_subtask,
-        color_camera_info5_subtask,
-        depth_camera_info5_subtask,
-    ]
-)
-
-######### MANIPULATION ########################################################################################
-move_to_point_ord_val_grab = OrderedCallsValidator(
-    subtasks=[move_to_point_subtask_grab]
-)
-move_to_point_ord_val_drop = OrderedCallsValidator(
-    subtasks=[move_to_point_subtask_drop]
-)
-
-pub_HRIMessage_text_ord_val = OrderedCallsValidator(
-    subtasks=[pub_HRIMessage_text_subtask]
-)
-
-list_topic_get_interface_publish_ord_val = OrderedCallsValidator(
-    subtasks=[
-        get_topics_subtask,
-        get_tohuman_interface_subtask,
-        pub_HRIMessage_text_subtask,
-    ]
-)
-
-ret_true_ord_val = OrderedCallsValidator(subtasks=[return_true_subtask])
-ret_false_ord_val = OrderedCallsValidator(subtasks=[return_false_subtask])
-
-start_navigate_action_ord_val = OrderedCallsValidator(
-    subtasks=[start_nav_action_subtask]
-)
-start_spin_action_ord_val = OrderedCallsValidator(subtasks=[start_spin_action_subtask])
-move_ahead_ord_val = OrderedCallsValidator(subtasks=[start_move_front_action_subtask])
-
-
-######### TASKS ############################################################################################
-def get_basic_tasks(
-    extra_tool_calls: int = 0,
-    prompt_detail: List[Literal["brief", "moderate", "descriptive"]] = [
-        "brief",
-        "moderate",
-        "descriptive",
-    ],
-    n_shots: List[Literal[0, 2, 5]] = [0, 2, 5],
-) -> List[Task]:
-    tasks: List[Task] = []
-
-    # Generate all combinations of prompt_detail and n_shots
-    for detail in prompt_detail:
-        for shots in n_shots:
-            task_args = TaskArgs(
-                extra_tool_calls=extra_tool_calls,
-                prompt_detail=detail,
-                examples_in_system_prompt=shots,
-            )
-
-            tasks.extend(
-                [
-                    GetROS2RGBCameraTask(
-                        validators=[color_image_ord_val],
-                        task_args=task_args,
-                    ),
-                    GetROS2TopicsTask(
-                        validators=[topics_ord_val],
-                        task_args=task_args,
-                    ),
-                    GetROS2DepthCameraTask(
-                        validators=[depth_image_ord_val],
-                        task_args=task_args,
-                    ),
-                    GetAllROS2CamerasTask(
-                        validators=[all_camera_images_notord_val],
-                        task_args=task_args,
-                    ),
-                    GetPointcloudTask(
-                        validators=[get_pointcloud_ord_val], task_args=task_args
-                    ),
-                    GetRobotDescriptionTask(
-                        validators=[get_robot_desc_ord_val], task_args=task_args
-                    ),
-                    CheckRobotHealthTask(
-                        validators=[
-                            diagnostics_ord_val,
-                            rosout_ord_val,
-                            joint_states_ord_val,
-                        ],
-                        task_args=task_args,
-                    ),
-                    AssessSensorDataQualityTask(
-                        validators=[sensor_data_val],
-                        task_args=task_args,
-                    ),
-                ]
-            )
-
-    return tasks
-
-
-def get_navigation_tasks(
-    extra_tool_calls: int = 0,
-    prompt_detail: List[Literal["brief", "moderate", "descriptive"]] = [
-        "brief",
-        "moderate",
-        "descriptive",
-    ],
-    n_shots: List[Literal[0, 2, 5]] = [0, 2, 5],
-) -> List[Task]:
-    tasks: List[Task] = []
-
-    for detail in prompt_detail:
-        for shots in n_shots:
-            task_args = TaskArgs(
-                extra_tool_calls=extra_tool_calls,
-                prompt_detail=detail,
-                examples_in_system_prompt=shots,
-            )
-            tasks.extend(
-                [
-                    NavigateToPointTask(
-                        validators=[start_navigate_action_ord_val],
-                        task_args=task_args,
-                    ),
-                    SpinAroundTask(
-                        validators=[start_spin_action_ord_val],
-                        task_args=task_args,
-                    ),
-                    MoveToBedTask(
-                        validators=[move_ahead_ord_val],
-                        task_args=task_args,
-                    ),
-                    MoveToFrontTask(
-                        validators=[move_ahead_ord_val],
-                        task_args=task_args,
-                    ),
-                ]
-            )
-
-    return tasks
-
-
-def get_manipulation_tasks(
-    extra_tool_calls: int = 0,
-    prompt_detail: List[Literal["brief", "moderate", "descriptive"]] = [
-        "brief",
-        "moderate",
-        "descriptive",
-    ],
-    n_shots: List[Literal[0, 2, 5]] = [0, 2, 5],
-) -> List[Task]:
-    tasks: List[Task] = []
-
-    for detail in prompt_detail:
-        for shots in n_shots:
-            task_args = TaskArgs(
-                extra_tool_calls=extra_tool_calls,
-                prompt_detail=detail,
-                examples_in_system_prompt=shots,
-            )
-            tasks.extend(
-                [
-                    MoveToPointTask(
-                        move_to_tool_input=MoveToPointToolInput(
-                            x=1.0, y=2.0, z=3.0, task="grab"
-                        ),
-                        validators=[move_to_point_ord_val_grab],
-                        task_args=task_args,
-                    ),
-                    MoveToPointTask(
-                        move_to_tool_input=MoveToPointToolInput(
-                            x=1.2, y=2.3, z=3.4, task="drop"
-                        ),
-                        validators=[move_to_point_ord_val_drop],
-                        task_args=task_args,
-                    ),
-                ]
-            )
-
-    return tasks
-
-
-def get_custom_interfaces_tasks(
-    extra_tool_calls: int = 0,
-    prompt_detail: List[Literal["brief", "moderate", "descriptive"]] = [
-        "brief",
-        "moderate",
-        "descriptive",
-    ],
-    n_shots: List[Literal[0, 2, 5]] = [0, 2, 5],
-) -> List[Task]:
-    tasks: List[Task] = []
-
-    for detail in prompt_detail:
-        for shots in n_shots:
-            task_args = TaskArgs(
-                extra_tool_calls=extra_tool_calls,
-                prompt_detail=detail,
-                examples_in_system_prompt=shots,
-            )
-            tasks.extend(
-                [
-                    PublishROS2HRIMessageTextTask(
-                        topic="/to_human",
-                        validators=[pub_HRIMessage_text_ord_val],
-                        task_args=task_args,
-                        text="Hello!",
-                    ),
-                    PublishROS2HRIMessageTextTask(
-                        topic="/to_human",
-                        validators=[list_topic_get_interface_publish_ord_val],
-                        task_args=task_args,
-                        text="Hello!",
-                    ),
-                ]
-            )
-
-    return tasks
-
-
-def get_spatial_tasks(
-    extra_tool_calls: int = 0,
-    prompt_detail: List[Literal["brief", "moderate", "descriptive"]] = [
-        "brief",
-        "moderate",
-        "descriptive",
-    ],
-    n_shots: List[Literal[0, 2, 5]] = [0, 2, 5],
-) -> Sequence[Task]:
-    tasks: List[Task] = []
-
-    # Categorize tasks by complexity based on question difficulty
-    easy_true_inputs = [
-        # Single object presence/detection
-        BoolImageTaskInput(
-            question="Is the light on in the room?",
-            images_paths=[IMG_PATH + "image_2.jpg"],
-        ),
-        BoolImageTaskInput(
-            question="Do you see the plant?", images_paths=[IMG_PATH + "image_2.jpg"]
-        ),
-        BoolImageTaskInput(
-            question="Are there any pictures on the wall?",
-            images_paths=[IMG_PATH + "image_3.jpg"],
-        ),
-        BoolImageTaskInput(
-            question="Is there a pillow on the armchain?",
-            images_paths=[IMG_PATH + "image_7.jpg"],
-        ),
-    ]
-
-    medium_true_inputs = [
-        # Object state or counting
-        BoolImageTaskInput(
-            question="Are there 3 pictures on the wall?",
-            images_paths=[IMG_PATH + "image_4.jpg"],
-        ),
-    ]
-
-    hard_true_inputs = [
-        # Spatial relationships between objects
-        BoolImageTaskInput(
-            question="Is the door on the left from the desk?",
-            images_paths=[IMG_PATH + "image_1.jpg"],
-        ),
-        BoolImageTaskInput(
-            question="Is there a plant behind the rack?",
-            images_paths=[IMG_PATH + "image_5.jpg"],
-        ),
-    ]
-
-    easy_false_inputs = [
-        # Single object presence/detection
-        BoolImageTaskInput(
-            question="Is someone in the room?", images_paths=[IMG_PATH + "image_1.jpg"]
-        ),
-        BoolImageTaskInput(
-            question="Do you see the plant?", images_paths=[IMG_PATH + "image_3.jpg"]
-        ),
-        BoolImageTaskInput(
-            question="Is there a red pillow on the armchair?",
-            images_paths=[IMG_PATH + "image_7.jpg"],
-        ),
-    ]
-
-    medium_false_inputs = [
-        # Object state or counting
-        BoolImageTaskInput(
-            question="Is the door open?", images_paths=[IMG_PATH + "image_1.jpg"]
-        ),
-        BoolImageTaskInput(
-            question="Are there 4 pictures on the wall?",
-            images_paths=[IMG_PATH + "image_4.jpg"],
-        ),
-    ]
-
-    hard_false_inputs = [
-        # Spatial relationships between objects
-        BoolImageTaskInput(
-            question="Is there a rack on the left from the sofa?",
-            images_paths=[IMG_PATH + "image_4.jpg"],
-        ),
-        BoolImageTaskInput(
-            question="Is there a plant on the right from the window?",
-            images_paths=[IMG_PATH + "image_6.jpg"],
-        ),
-    ]
-
-    for detail in prompt_detail:
-        for shots in n_shots:
-            task_args = TaskArgs(
-                extra_tool_calls=extra_tool_calls,
-                prompt_detail=detail,
-                examples_in_system_prompt=shots,
-            )
-
-            [
-                BoolImageTaskEasy(
-                    task_input=input_item,
-                    validators=[ret_true_ord_val],
-                    task_args=task_args,
-                )
-                for input_item in easy_true_inputs
-            ]
-
-            tasks.extend(
-                [
-                    BoolImageTaskEasy(
-                        task_input=input_item,
-                        validators=[ret_true_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in easy_true_inputs
-                ]
-            )
-
-            tasks.extend(
-                [
-                    BoolImageTaskEasy(
-                        task_input=input_item,
-                        validators=[ret_false_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in easy_false_inputs
-                ]
-            )
-
-            tasks.extend(
-                [
-                    BoolImageTaskMedium(
-                        task_input=input_item,
-                        validators=[ret_true_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in medium_true_inputs
-                ]
-            )
-
-            tasks.extend(
-                [
-                    BoolImageTaskMedium(
-                        task_input=input_item,
-                        validators=[ret_false_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in medium_false_inputs
-                ]
-            )
-
-            tasks.extend(
-                [
-                    BoolImageTaskHard(
-                        task_input=input_item,
-                        validators=[ret_true_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in hard_true_inputs
-                ]
-            )
-
-            tasks.extend(
-                [
-                    BoolImageTaskHard(
-                        task_input=input_item,
-                        validators=[ret_false_ord_val],
-                        task_args=task_args,
-                    )
-                    for input_item in hard_false_inputs
-                ]
-            )
-
-    return tasks
 
 
 def get_tasks(
